pyproject.toml
[tool.poetry]
name = "menusuncookies"
version = "0.1.0"
description = ""
authors = ["Gabriel Aguiar Reis <lugafeagre@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
markdown = "^3.5.1"
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.8"
mkdocs-mermaid2-plugin = "^1.1.1"
pymdown-extensions = "^10.4"
python = "^3.12"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
pip-audit = "^2.6.1"
pytest = "^7.4.3"
pylint = "^3.0.2"
pytest-cov = "^4.1.0"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
